feat(interfaces): extract SmartAccountPlugin and SmartAccountPolicy into new reusable crate
- New crate: crates/smart-account-interfaces
- Move plugin and policy traits (with contractclient) to shared crate
- Re-export clients in crate root for downstream imports
- Refactor smart-account and deny-list-policy to depend on shared traits
- Keep SmartAccountInterface internal to avoid coupling to implementation types

Includes workspace and Cargo updates. All tests pass locally (fmt/clippy/test).
