+++ crates/smart-account-interfaces/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "smart-account-interfaces"
+version = "0.1.0"
+edition = "2021"
+license = "Apache-2.0"
+repository = "https://github.com/Crossmint/stellar-smart-account"
+description = "Reusable interfaces (traits) for Stellar Smart Account plugins and policies."
+
+[lib]
+name = "smart_account_interfaces"
+path = "src/lib.rs"
+
+[dependencies]
+soroban-sdk = "22.0.0"
