+++ crates/smart-account-interfaces/src/lib.rs
@@ -1,0 +1,5 @@
+pub mod plugin;
+pub mod policy;
+
+pub use plugin::{SmartAccountPlugin, SmartAccountPluginClient};
+pub use policy::{SmartAccountPolicy, SmartAccountPolicyClient};
